chore(housekeeping): avoid imports in helpers.js
